Move the source-impact-report CLI to src/tools/, delete src/cli/
source-impact-report already had the shape this reorganization is aiming
for: a 621-line argparse frontend over ~1,600 lines of library code in
src/model/ and src/reports/. It was only in the wrong directory, and it
was the sole occupant of src/cli/.

Its src.* imports are untouched. The pipeline reaches this tool through
the `uv run source-impact-report` shell command in reports.snakefile
rather than by importing it, so only the console-script target moves.

tests/cli/source_impact_report/ becomes tests/tools/source_impact_report/,
and tests/tools/ gains the __init__.py that was previously forbidden.
